Removed most of the messy conduits.
My opinion on Conduits: Great for a specific task, writing server. Not
very good at implementing the internals of a simple API wrapper.  By
requiring the use of conduits, the consumer of this library is forced
to use a number of constructs and dependencies which may (were, in my
case) not be necessary, or desirable.

This doesn't resemble the original library in any way, as it is now
just a constrained HTTP request builder, but at least all the API is
described, and the implementation of the Rest portion, excluding
actual request and response handling (which can, if desired, be
wrapped trivially in conduits) is completely pure now.